move and create enum
