+++ functions/addBook.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { addBook } from "@/actions/bookActions";
-import { Status } from "@/types/statusEnum";
+import { Status } from "@/types/enums/statusEnum";
 import { currentUser } from "@clerk/nextjs/server";
 
 export const addBookToList = async (
