+++ functions/addBookNote.ts
@@ -3,7 +3,7 @@
 import { addBook } from "@/actions/bookActions";
 import { addNote } from "@/actions/noteActions";
 import { InsertBookNote } from "@/lib/schema";
-import { Status } from "@/types/statusEnum";
+import { Status } from "@/types/enums/statusEnum";
 import { currentUser } from "@clerk/nextjs/server";
 
 export const addBookNote = async (
