+++ types/enums/goalsEnum.ts
@@ -1,0 +1,5 @@
+export enum Status {
+  Month,
+  Year,
+  Quarter,
+}
